+++ tests/test_jinja.py
@@ -131,6 +131,41 @@ def test_success_outline_complete():
     assert result == expected
 
 
+def test_success_outline_size_none():
+    env = make_environment('{{ heroicon_outline("academic-cap", size=None) }}')
+    template = env.get_template("index")
+
+    result = template.render()
+
+    expected_py37 = (
+        '<svg aria-hidden="true" fill="none" stroke="currentColor"'
+        ' stroke-width="1.5" viewBox="0 0 24 24">\n  <path d="M4.26'
+        " 10.147a60.436 60.436 0 00-.491 6.347A48.627 48.627 0 0112 20.904a48.627"
+        " 48.627 0 018.232-4.41 60.46 60.46 0 00-.491-6.347m-15.482 0a50.57 50.57 0"
+        " 00-2.658-.813A59.905 59.905 0 0112 3.493a59.902 59.902 0 0110.399"
+        " 5.84c-.896.248-1.783.52-2.658.814m-15.482 0A50.697 50.697 0 0112"
+        " 13.489a50.702 50.702 0 017.74-3.342M6.75 15a.75.75 0 100-1.5.75.75 0 000"
+        " 1.5zm0 0v-3.675A55.378 55.378 0 0112 8.443m-7.007 11.55A5.981 5.981 0 006.75"
+        ' 15.75v-1.5" stroke-linecap="round" stroke-linejoin="round" />\n</svg>'
+    )
+    expected_py38plus = (
+        '<svg fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor"'
+        ' aria-hidden="true">\n  <path stroke-linecap="round"'
+        ' stroke-linejoin="round" d="M4.26 10.147a60.436 60.436 0 00-.491 6.347A48.627'
+        " 48.627 0 0112 20.904a48.627 48.627 0 018.232-4.41 60.46 60.46 0"
+        " 00-.491-6.347m-15.482 0a50.57 50.57 0 00-2.658-.813A59.905 59.905 0 0112"
+        " 3.493a59.902 59.902 0 0110.399 5.84c-.896.248-1.783.52-2.658.814m-15.482"
+        " 0A50.697 50.697 0 0112 13.489a50.702 50.702 0 017.74-3.342M6.75 15a.75.75 0"
+        " 100-1.5.75.75 0 000 1.5zm0 0v-3.675A55.378 55.378 0 0112 8.443m-7.007"
+        ' 11.55A5.981 5.981 0 006.75 15.75v-1.5" />\n</svg>'
+    )
+    if sys.version_info < (3, 8):
+        expected = expected_py37
+    else:
+        expected = expected_py38plus
+    assert result == expected
+
+
 def test_success_solid():
     env = make_environment(
         '{{ heroicon_solid("academic-cap", size=40, class="h-4 w-4 inline") }}'