+++ HISTORY.rst
@@ -2,6 +2,8 @@
 History
 =======
 
+* Support ``size=None`` to disable setting the ``width`` and ``height`` icon attributes.
+
 2.0.0 (2022-10-05)
 ------------------